chore(eventarc): Refactor eventarc error logic
